requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# chatbot/app.py: 1
Flask == 1.1.2

# chatbot/fb_graph_api/send_api.py: 4
python_magic == 0.4.24

# chatbot/fb_graph_api/send_api.py: 2
requests == 2.25.1

# chatbot/fb_graph_api/send_api.py: 3
requests_toolbelt == 0.9.1

# for the wsgi server
gunicorn == 20.1.0

